rename util dir to dataloader
